refactor(accounts,courses): extract new courses app and move models/migrations
- Add new app `courses` (admin, apps, enums, models, views, tests)
- Move course-related code/migrations from `home` -> `courses`
- Wire signals in accounts and update settings
- Update tests and fix imports
