+++ uclases/core/settings.py
@@ -34,8 +34,9 @@ INSTALLED_APPS = [
 	'django.contrib.sessions',
 	'django.contrib.messages',
 	'django.contrib.staticfiles',
-    'accounts',
+    'accounts.apps.AccountsConfig',
 	'home',
+    'courses',
 ]
 
 MIDDLEWARE = [
